add score board

+++ setupgit.sh
@@ -9,7 +9,7 @@ cd nasa-judge-api
 
 for a in $(ls -a)
 do
-    if [ "$a" != "." ] && [ "$a" != ".." ] && [ "$a" != ".git" ] && [ "$a" != "README.md" ] && [ "$a" != "install.sh" ] && [ "$a" != "remove.sh" ] && [ "$a" != "setupdefaultiptables.sh" ] && [ "$a" != "iptablesconf.conf" ] && [ "$a" != "setupgit.sh" ]
+    if [ "$a" != "." ] && [ "$a" != ".." ] && [ "$a" != ".git" ] && [ "$a" != "README.md" ] && [ "$a" != "install.sh" ] && [ "$a" != "remove.sh" ] && [ "$a" != "setupdefaultiptables.sh" ] && [ "$a" != "iptablesconf.conf" ] && [ "$a" != "setupgit.sh" ] && [ "$a" != "db.conf" ]
     then
         rm -rf $a
     fi
@@ -19,12 +19,13 @@ for a in $(ls -a /etc/nasajudgeapi)
 do
     if [ "$a" != "." ] && [ "$a" != ".." ] && [ "$(cat /etc/nasajudgeapi/.gitignore | sed 's/\/.*//g' | sed '/^!.*/d' | grep -P "^$(echo "$a" | sed 's/\./\\\./g')$")" == "" ]
     then
-        cp -r /etc/nasajudgeapi/$a $a
+        sudo cp -r /etc/nasajudgeapi/$a $a
     fi
 done
 
-cp -r /etc/wireguard/testserverfirewall.sh testserverfirewall.sh
+sudo cp -r /etc/wireguard/testserverfirewall.sh testserverfirewall.sh
 
 > node.conf
 
-cp /etc/systemd/system/nasajudgeapi.service nasajudgeapi.service
+sudo cp /etc/systemd/system/nasajudgeapi.service nasajudgeapi.service
+sudo cp /etc/systemd/system/nasajudgesql.service nasajudgesql.service
